+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This changelog also contains important changes in dependencies.
 ## [Unreleased]
 ### Fixed
 - Absolute transform propagation during `use` resolving.
+- Absolute transform propagation during nested `svg` resolving.
 - `Node::abs_transform` documentation. The current element's transform _is_ included.
 
 ## [0.40.0] - 2024-02-17